LoginToken/backend/requirements.txt
pip
wheel
fastapi[all]
uvicorn[standard]
gunicorn
SQLAlchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
pyjwt
